% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{dmix}
\alias{dmix}
\title{Mixture Density Function}
\usage{
dmix(x, d1, s1, d2, s2, w, log = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{d1, d2}{means}

\item{s1, s2}{standard deviations}

\item{w}{mixing weight (0 < w < 1)}

\item{log}{logical}
}
\value{
numeric vector
}
\description{
Density of a 2-component normal mixture
}
