% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab1_Box.R
\name{Lab1_Box}
\alias{Lab1_Box}
\title{Create a boxplot}
\usage{
Lab1_Box(values)
}
\arguments{
\item{values}{Create a vector of values}
}
\value{
Boxplot of value given with color blue
}
\description{
Create a boxplot
}
