% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab1_QuickBox.R
\name{Lab1_QuickBox}
\alias{Lab1_QuickBox}
\title{Create a boxplot}
\usage{
Lab1_QuickBox(values)
}
\arguments{
\item{values}{Create a vector of values}
}
\value{
Boxplot
}
\description{
Create a boxplot
}
