% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab1_QuickPie.R
\name{Lab1_QuickPie}
\alias{Lab1_QuickPie}
\title{Create a Pie Plot}
\usage{
Lab1_QuickPie(value)
}
\arguments{
\item{value}{A numeric vector of values for Pie Chart}
}
\value{
Pie Plot
}
\description{
Create a Pie Plot
}
\examples{
\dontrun {# Pie Chart of Fish SPECIES pie(tabFish)}
}
