% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab2_zScore.R
\name{Lab2_zScore}
\alias{Lab2_zScore}
\title{Find the z-score}
\usage{
Lab2_zScore(values)
}
\arguments{
\item{values}{A list of vectors of values}
}
\value{
z-score vector
}
\description{
Find the z-score
}
