% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab3_Plot.R
\name{Lab3_Plot}
\alias{Lab3_Plot}
\title{Creates a plot with blue dots}
\usage{
Lab3_Plot(xVariable, yVariable, xlabel, ylabal, title, datatype)
}
\arguments{
\item{title}{title of the plot}

\item{x}{x-axis variables}

\item{y}{y-axix variables}

\item{xlab}{x label}

\item{ylab}{y label}

\item{dataType}{dataset being called in}
}
\value{
returns the plot
}
\description{
Creates a plot with blue dots
}
