% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab4_ResidvsFittedPlot.R
\name{Lab4_ResidvsFittedPlot}
\alias{Lab4_ResidvsFittedPlot}
\title{Plot Residual vs Fitted Value}
\usage{
Lab4_ResidvsFittedPlot(x, y, data)
}
\arguments{
\item{x}{x Variable}

\item{y}{y Variable}

\item{data}{Linear model temp location}
}
\value{
Residual vs Fitted Values Plot
}
\description{
Plot Residual vs Fitted Value
}
