% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab5_mynbin.R
\name{Lab5_mynbin}
\alias{Lab5_mynbin}
\title{Negative Binomial Distribution}
\usage{
Lab5_mynbin(y, r, p)
}
\arguments{
\item{y}{Number of fails}

\item{r}{The target number of successes}

\item{p}{Probability of Success in trials}
}
\value{
A decimal number of a probability
}
\description{
Negative Binomial Distribution
}
\examples{
Lab5_mynbin(y = 5, r = 3, p = 0.5)
}
