% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab6_myncurve.R
\name{Lab6_myncurve}
\alias{Lab6_myncurve}
\title{Creates density plot using dnorm when given mean and standard deviation}
\usage{
Lab6_myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{cutoff value}
}
\value{
Returns density plot with area colored in
}
\description{
Creates density plot using dnorm when given mean and standard deviation
}
\examples{
Lab6_myncurve(mu = 0, sigma = 1, a = 1)
}
