% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab8_mycltu.R
\name{Lab8_mycltu}
\alias{Lab8_mycltu}
\title{Central Limit Theorem for Uniform Distribution}
\usage{
Lab8_mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{Sample Size}

\item{iter}{Number of iterations}

\item{a}{Lower bound of Uniform Distribution}

\item{b}{Upper bound of Uniform Distribution}
}
\value{
Output histogram bars with the distribution of samples
}
\description{
Central Limit Theorem for Uniform Distribution
}
\examples{
Lab8_mycltu(n = 10, iter = 1000)
}
