% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATH4753S26CODY-package.R
\docType{package}
\name{MATH4753S26CODY-package}
\alias{MATH4753S26CODY}
\alias{MATH4753S26CODY-package}
\title{MATH4753S26CODY Package}
\description{
Functions and tools for MATH 4753 labs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cchan555/MATH4753S26CODY}
  \item Report bugs at \url{https://github.com/cchan555/MATH4753S26CODY/issues}
}

}
\author{
\strong{Maintainer}: Cody Chan \email{cody.l.chan-1@ou.edu}

}
