% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab9_myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Creates Bootstrap Density plot with estimated intervals}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of samples being reused}

\item{x}{sample data}

\item{fun}{the statistics being used}

\item{alpha}{the confidence interval}

\item{cx}{the graph labels (scaling)}

\item{...}{other parameters if needed}
}
\value{
returns the density plot with confidence intervals
}
\description{
Creates Bootstrap Density plot with estimated intervals
}
