% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysd.R
\name{mysd}
\alias{mysd}
\title{Title My Standard Deviation Function}
\usage{
mysd(x, na.rm = TRUE)
}
\arguments{
\item{x}{A quantitative vector}

\item{na.rm}{A Logical indicating whether to remove NA value before computation}
}
\value{
A quantitative vector with square (SD of the numeric vector)
}
\description{
Title My Standard Deviation Function
}
\examples{
mysd(c(1, 2, 3, 4, 5))
}
