% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project1_ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Overbooking Problem (Number of tickets to sell a flight)}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{N - Number of seats on the plane}

\item{gamma}{Gamma - acceptable probability for overbooking}

\item{p}{p - probability that a ticket of a person will show up}
}
\value{
returns Discrete and Continuous Plots of the overbooking problem
}
\description{
Overbooking Problem (Number of tickets to sell a flight)
}
