% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{pmix}
\alias{pmix}
\title{Mixture CDF}
\usage{
pmix(q, d1, s1, d2, s2, w)
}
\arguments{
\item{q}{quantiles}

\item{d1, d2}{means}

\item{s1, s2}{standard deviations}

\item{w}{mixing weight (0 < w < 1)}
}
\description{
Mixture CDF
}
