% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{rmix}
\alias{rmix}
\title{Random Generation}
\usage{
rmix(n, d1, s1, d2, s2, w)
}
\arguments{
\item{n}{sample size}

\item{d1, d2}{means}

\item{s1, s2}{standard deviations}

\item{w}{mixing weight (0 < w < 1)}
}
\description{
Random Generation
}
