% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab10_mymlnorm.R
\name{Lab10_mymlnorm}
\alias{Lab10_mymlnorm}
\title{Maximum Likelihood for Normal Distribution of two parameters}
\usage{
Lab10_mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{The sample size}

\item{mu}{The mean value}

\item{sig}{The standard deviation}

\item{...}{other arguments from 'contour()'}
}
\value{
A list of Matrix coordinates of maximum likelihood and plot
}
\description{
Maximum Likelihood for Normal Distribution of two parameters
}
\examples{
Lab10_mymlnorm(x=c(10,12,13,15,12, 11, 10),mu=seq(10,15,length=1000),sig=seq(0.1,4,length=1000))
}
