% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab3_Plot.R
\name{Lab3_Plot}
\alias{Lab3_Plot}
\title{Creates a plot with blue dots with limits}
\usage{
Lab3_Plot(xVariable, yVariable, xlabel, ylabel, title)
}
\arguments{
\item{xVariable}{x-axis variables}

\item{yVariable}{y-axix variables}

\item{xlabel}{x label}

\item{ylabel}{y label}

\item{title}{title of the plot}
}
\value{
returns the plot
}
\description{
Creates a plot with blue dots with limits
}
