% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysd.R
\name{mysd}
\alias{mysd}
\title{Title}
\usage{
mysd(x, na.rm = TRUE)
}
\arguments{
\item{x}{A quantitative vector}

\item{na.rm}{}
}
\value{
A quantitative vector with squre
}
\description{
Title
}
\examples{
\dontrun mysq(1:10)
}
